LayoutTests/fast/dom/HTMLDialogElement/dialog-enabled.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.internals)
    internals.settings.setDialogElementEnabled(true);
</script>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<dialog id="mydialog"></dialog>
<script>
description("The DIALOG element should be recognized if the dialog feature is on.");
dialog = document.getElementById('mydialog');
shouldBeTrue("typeof dialog.show == 'function'");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
